epwa2708it6/anysell/requirements.txt
Flask==1.0.2
requests==2.21.0
line-bot-sdk
dialogflow==0.4.0
apiai==1.2.3
jsondatabase==0.1.7
numpy==1.13.3
urllib3==1.22
python-dotenv==0.8.2
pusher==2.0.1
gspread
oauth2client
mysqlclient
pymysql
gunicorn==19.6.0
firebase_admin
Pillow
Flask-Login
flask-session
psycopg2
Flask-Migrate
Flask-Script
flask_uploads
flask-wtf
Werkzeug